requirements.txt
discord.py==2.3.2
python-dotenv==0.20.0